Avancer

+++ src/Controller/MagasinController.php
@@ -12,15 +12,28 @@ use Symfony\Component\Routing\Annotation\Route;
 class MagasinController extends AbstractController
 {
     /**
-     * @Route("/", name="magasin_list")
+     * @Route("/magasin", name="magasin_list")
      */
     public function index(): Response
     {
         $em = $this->getDoctrine()->getManager();
         $magasins = $em->getRepository(Magasin::class)->findAll();
 
+        $admin = false;
+        $user = $this->getUser();
+        if($user != null){
+            $roles = $user->getRoles();
+            foreach ($roles as $role){
+                if($role == "ROLE_ADMIN"){
+                    $admin = true;
+                    break;
+                }
+            }
+        }
+
         return $this->render('magasin/index.html.twig', [
-            'magasins' => $magasins
+            'magasins' => $magasins,
+            'admin' => $admin,
         ]);
     }
 
@@ -32,8 +45,21 @@ class MagasinController extends AbstractController
     public function detailMagasin(Magasin $magasin){
         $em = $this->getDoctrine()->getManager();
 
+        $admin = false;
+        $user = $this->getUser();
+        if($user != null){
+            $roles = $user->getRoles();
+            foreach ($roles as $role){
+                if($role == "ROLE_ADMIN"){
+                    $admin = true;
+                    break;
+                }
+            }
+        }
+
         return $this->render('magasin/detail.html.twig', [
-            'magasin' => $magasin
+            'magasin' => $magasin,
+            'admin' => $admin,
         ]);
     }
 
@@ -43,6 +69,21 @@ class MagasinController extends AbstractController
      * @return Response
      */
     public function addMagasin(Request $request){
+        $admin = false;
+        $user = $this->getUser();
+        if($user != null){
+            $roles = $user->getRoles();
+            foreach ($roles as $role){
+                if($role == "ROLE_ADMIN"){
+                    $admin = true;
+                    break;
+                }
+            }
+        }
+        if($admin == false){
+            return $this->redirectToRoute("magasin_list");
+        }
+
         $em = $this->getDoctrine()->getManager();
         $addFrom = $this->createForm(MagasinType::class);
         $addFrom->handleRequest($request);
@@ -62,6 +103,20 @@ class MagasinController extends AbstractController
      * @Route("magasin/edit/{id}", name="magasin_edit")
      */
     public function editMagasin(Request $request, Magasin $magasin){
+        $admin = false;
+        $user = $this->getUser();
+        if($user != null){
+            $roles = $user->getRoles();
+            foreach ($roles as $role){
+                if($role == "ROLE_ADMIN"){
+                    $admin = true;
+                    break;
+                }
+            }
+        }
+        if($admin == false){
+            return $this->redirectToRoute("magasin_list");
+        }
         $em = $this->getDoctrine()->getManager();
 
         $editForm = $this->createForm(MagasinType::class, $magasin);
@@ -81,8 +136,23 @@ class MagasinController extends AbstractController
     /**
      * @Route("magasin/delete/{id}", name="magasin_delete")
      * @param Magasin $magasin
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
     public function deleteMagasin(Magasin $magasin){
+        $admin = false;
+        $user = $this->getUser();
+        if($user != null){
+            $roles = $user->getRoles();
+            foreach ($roles as $role){
+                if($role == "ROLE_ADMIN"){
+                    $admin = true;
+                    break;
+                }
+            }
+        }
+        if($admin == false){
+            return $this->redirectToRoute("magasin_list");
+        }
         $em = $this->getDoctrine()->getManager();
         $em->remove($magasin);
         $em->flush();
